AI/TextCtrl: database ctrls

+++ uppsrc/AI/TextCtrl/TextDesigner.cpp
@@ -137,7 +137,7 @@ void ToolAppCtrl::MakeComponentParts(ArrayCtrl& parts) {
 }
 
 void ToolAppCtrl::Load(const String& includes, const String& filename, Stream& in, byte charset) {
-	data = in.Get(in.GetSize());
+	data = in.Get((int)in.GetSize());
 	OnLoad(data, filename);
 	data_filepath = filename;
 }
